+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.6.0] - 2020-10-17
+
+### Added
+- Revised installation paths.
+- Fall back on PATH when searching for installations.
+- Check for successful MonoInstallationFinder reflection.
+- Export FrameworkPathOverride setting for integrations that allow passing environment variables.
+- Native opening for editors that require it.
+- Custom arguments per editor.
+- Generate project files on demand.
+- Discovery APIs to find and use installations.
+- Support for BBEdit and CotEditor on macOS.
+
+### Removed
+- Launcher library and its APIs.
+- Setting class.
+- OmniSharp reloading.
+- Matching compiler version on 2020.2+ (Unity fixed it).
+
 ## [0.5.1] - 2020-10-12
 
 ### Added
@@ -8,7 +27,7 @@
 - `package.json` EditorConfig rules.
 - .NET EditorConfig rules.
 
-### Changed
+### Removed
 - Don't include the README in the package.
 
 ## [0.5.0] - 2020-10-06
@@ -85,7 +104,7 @@
 - Support for MacVim on macOS.
 - Support for Sublime Text 3 on macOS.
 
-[Unreleased]: https://github.com/frarees/easyeditor/compare/v0.5.1...HEAD
+[0.6.0]: https://github.com/frarees/easyeditor/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/frarees/easyeditor/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/frarees/easyeditor/compare/v0.4.3...v0.5.0
 [0.4.3]: https://github.com/frarees/easyeditor/compare/v0.4.2...v0.4.3